+++ modules/package.nix
@@ -136,11 +136,6 @@ in {
         type = bool;
         default = false;
       };
-
-      ghcOptions = mkOption {
-        type = nullOr str;
-        default = null;
-      };
     };
 
     components = let