+++ modules/plan.nix
@@ -229,6 +229,11 @@ let
       type = listOfFilteringNulls str;
       default = (def.hardeningDisable or []);
     };
+    ghcOptions = mkOption {
+      type = listOfFilteringNulls str;
+      default = def.ghcOptions or [];
+      merge = a: b: __trace "HELLO";
+    };
   };